backend/requirements.txt
alembic==1.6.5
bcrypt==3.2.0
cffi==1.14.5
click==8.0.1
Flask==1.1.2
Flask-Migrate==3.0.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
importlib-metadata==4.6.0
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.4
MarkupSafe==2.0.1
pkg-resources==0.0.0
pycparser==2.20
python-dateutil==2.8.1
python-dotenv==0.18.0
python-editor==1.0.4
six==1.16.0
SQLAlchemy==1.4.20
typing-extensions==3.10.0.0
Werkzeug==2.0.1
zipp==3.4.1
